pyproject.toml
[project]
name = "cpthinkstruct"
version = "0.1.0"
description = "Local patent similarity search with Sentence-Transformers and FAISS (uv-managed)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.3.1",
    "faiss-cpu>=1.13.0",
    "numpy>=2.2.6",
    "orjson>=3.11.4",
    "pydantic>=2.12.4",
    "pyyaml>=6.0.3",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "tqdm>=4.67.1",
]

[tool.uv]
index-url = "https://pypi.org/simple"
